labelling-tool/requirements.txt
azure-storage-blob==12.14.1
azure-identity==1.7.0
dash==2.6.1
numpy==1.19.5
plotly==5.6.0
scikit-image==0.19.3
pandas==1.4.3
gunicorn==20.0.4
dash_bootstrap_components==1.2.1
redis==4.3.4
fastapi==0.74.0
fastapi-login==1.8.3
path==16.5.0
pyodbc==4.0.34
uvicorn==0.17.5
python-dotenv==0.21.0
python-multipart==0.0.5
azure-keyvault-secrets==4.6.0